Refactor rate_limiter into a simpler, module-level interface
