+++ setup.py
@@ -3,7 +3,7 @@ from setuptools import setup, find_packages, Extension
 setup(
     name="pyinstrument",
     packages=['pyinstrument'],
-    ext_modules=[Extension('rate_limiter', sources=['rate_limiter.c'])],
+    ext_modules=[Extension('_statistical_profiler', sources=['_statistical_profiler.c'])],
     version="0.13.1",
     description="A call stack profiler for Python. Inspired by Apple's Instruments.app",
     author='Joe Rickerby',
